+++ src/layers/MultiscaleImageLayer/MultiscaleImageLayer.js
@@ -20,6 +20,7 @@ const defaultProps = {
   lensRadius: { type: 'number', value: 100, compare: true },
   lensBorderColor: { type: 'array', value: [255, 255, 255], compare: true },
   lensBorderRadius: { type: 'number', value: 0.02, compare: true },
+  maxRequests: { type: 'number', value: 10, compare: true },
   onClick: { type: 'function', value: null, compare: true }
 };
 
@@ -43,6 +44,7 @@ const defaultProps = {
  * @param {number} props.lensRadius Pixel radius of the lens (default: 100).
  * @param {number} props.lensBorderColor RGB color of the border of the lens (default [255, 255, 255]).
  * @param {number} props.lensBorderRadius Percentage of the radius of the lens for a border (default 0.02).
+ * @param {number} props.maxRequests Maximum parallel ongoing requests allowed before aborting.
  * @param {number} props.onClick Hook function from deck.gl to handle clicked-on objects.
  */
 
@@ -79,12 +81,13 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       lensSelection,
       lensBorderColor,
       lensBorderRadius,
+      maxRequests,
       onClick
     } = this.props;
     const { tileSize, numLevels, dtype } = loader;
     const { unprojectLensBounds } = this.state;
     const noWebGl2 = !isWebGL2(this.context.gl);
-    const getTileData = async ({ x, y, z }) => {
+    const getTileData = async ({ x, y, z, signal }) => {
       const tile = await loader.getTile({
         x,
         y,
@@ -95,7 +98,8 @@ export default class MultiscaleImageLayer extends CompositeLayer {
         // The image-tile example works without, this but I have a feeling there is something
         // going on with our pyramids and/or rendering that is different.
         z: Math.round(-z + Math.log2(512 / tileSize)),
-        loaderSelection
+        loaderSelection,
+        signal
       });
       if (tile) {
         tile.data = noWebGl2 ? to32BitFloat(tile.data) : tile.data;
@@ -120,6 +124,7 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       colorValues,
       sliderValues,
       channelIsOn,
+      maxRequests,
       domain,
       // We want a no-overlap caching strategy with an opacity < 1 to prevent
       // multiple rendered sublayers (some of which have been cached) from overlapping