+++ src/loaders/zarrLoader.js
@@ -1,6 +1,9 @@
 import { BoundsCheckError } from 'zarr';
 
-import { guessRgb, padTileWithZeros } from './utils';
+import { guessRgb, padTileWithZeros, byteSwapInplace } from './utils';
+import { DTYPE_VALUES } from '../constants';
+import HTTPStore from './httpStore';
+
 /**
  * This class serves as a wrapper for fetching zarr data from a file server.
  * */
@@ -60,7 +63,8 @@ export default class ZarrLoader {
    * @param {Array} loaderSelection, Array of valid dimension selections
    * @returns {Object} data: TypedArray[], width: number (tileSize), height: number (tileSize)
    */
-  async getTile({ x, y, z, loaderSelection = [] }) {
+  async getTile({ x, y, z, loaderSelection = [], signal }) {
+    const { TypedArray } = DTYPE_VALUES[this.dtype];
     const source = this._getSource(z);
     const [xIndex, yIndex] = ['x', 'y'].map(k => this._dimIndices.get(k));
 
@@ -68,10 +72,28 @@ export default class ZarrLoader {
       const chunkKey = this._serializeSelection(sel);
       chunkKey[yIndex] = y;
       chunkKey[xIndex] = x;
-      const {
-        data,
-        shape: [height, width]
-      } = await source.getRawChunk(chunkKey);
+
+      const key = source.keyPrefix + chunkKey.join('.');
+      let buffer;
+      try {
+        buffer = await source.store.getItem(key, { signal });
+      } catch (err) {
+        if (err.name !== 'AbortError') {
+          throw err;
+        }
+        return null;
+      }
+      let bytes = new Uint8Array(buffer);
+      if (source.compressor) {
+        bytes = await source.compressor.decode(bytes);
+      }
+      const data = new TypedArray(bytes.buffer);
+      if (source.dtype[0] === '>') {
+        // big endian
+        byteSwapInplace(data);
+      }
+      const width = source.chunks[xIndex];
+      const height = source.chunks[yIndex];
       if (height < this.tileSize || width < this.tileSize) {
         return padTileWithZeros(
           { data, width, height },
@@ -81,8 +103,8 @@ export default class ZarrLoader {
       }
       return data;
     });
-
     const data = await Promise.all(dataRequests);
+    if (source.store instanceof HTTPStore && signal?.aborted) return null;
     return {
       data,
       width: this.tileSize,